refactor: move interfaces to app directory
