+++ src/interfaces/app/IEnv.d.ts
