+++ src/interfaces/app/INuxtRoute.d.ts
