+++ src/interfaces/dom/HTMLElementEvent.d.ts
